fix: refactored faq page code and file structure
