+++ Frontend/src/assets/svgs/index.js
@@ -1,5 +1,6 @@
 import {ReactComponent as ContactChatIcon} from './contactChat.svg'
 import {ReactComponent as ContactPhoneIcon} from './contact-phone-icon.svg'
 import { ReactComponent as ContactLocationIcon } from './Conatct-locaton-icon.svg'
+import { ReactComponent as FaqIcon } from './faqIcon.svg'
 
-export {ContactChatIcon, ContactPhoneIcon, ContactLocationIcon}
+export { ContactChatIcon, ContactPhoneIcon, ContactLocationIcon, FaqIcon };
